Running custom runtime in CDK
