+++ DeliveryApi/LambdaBaseContainerImage
@@ -1,6 +1,6 @@
 FROM public.ecr.aws/lambda/java:11
 
-COPY ApiHandlers/build/libs/shadow-base.jar ${LAMBDA_TASK_ROOT}/lib/
+COPY ApiHandlers/build/libs/uber-minimized.jar ${LAMBDA_TASK_ROOT}/lib/
 COPY ApiHandlers/resources/rds-ca-2019-root.pem ${LAMBDA_TASK_ROOT}/rds-ca-2019-root.pem
 
 CMD [ "com.ilmlf.delivery.api.handlers.CreateSlots::handleRequest"]
